chore: init monorepo (#2)
This pull request includes changes to the configuration settings and the
addition of a new package for the menu module. The most important
changes are summarized below:

Configuration settings:

*
[`.vscode/settings.json`](diffhunk://#diff-a5de3e5871ffcc383a2294845bd3df25d3eeff6c29ad46e3a396577c413bf357R1-R3):
Added a configuration to set the live server port to 5501.

New package for menu module:

*
[`modules/menu/package.json`](diffhunk://#diff-0877005e3c57f1bd90fd5d753727ba74fc6db7a7b3729ed2cdf850bad24e0032R1-R38):
Added a new package file for the menu module, including scripts for
development, build, linting, and preview, as well as dependencies and
devDependencies for React, Tailwind CSS, TypeScript, and other tools.

Co-authored-by: Mathieu Audebert <mathieu.adbrt@gmail.com>