feat(menu-mobile): increase menu icon size (#6)
This pull request includes a small change to the `Navbar` component in
the `modules/menu-mobile/src/components/navbar.tsx` file. The change
modifies the button and span elements to improve their visual
appearance.

*
[`modules/menu-mobile/src/components/navbar.tsx`](diffhunk://#diff-59c9b06d9b76c8a6dfba3428e502c5a37ac58fa339875ca082e949b9cff622f7L87-R93):
Updated the button to have a larger scale and increased the height of
the span elements for better visibility.